ingestion/requirements.txt
cyclonedx-python-lib==11.7.0
psycopg2==2.9.11
sqlalchemy==2.0.48
requests==2.33.1